Completed the product service, created order-service

+++ product-service/src/main/java/com/nonsysncbobbal/microservices/product_service/dto/ProductDTO.java
@@ -1,4 +1,6 @@
 package com.nonsysncbobbal.microservices.product_service.dto;
 
-public record ProductDTO(String name, String description, double price) {
+import java.math.BigDecimal;
+
+public record ProductDTO(String name, String description, BigDecimal price) {
 }
